creates abstract tests to MapperClass
