+++ sdk/src/test/java/com/microsoft/azure/cosmosdb/mapper/MapperTest.java
@@ -22,322 +22,13 @@
  */
 package com.microsoft.azure.cosmosdb.mapper;
 
-import com.microsoft.azure.cosmosdb.ConnectionPolicy;
-import com.microsoft.azure.cosmosdb.ConsistencyLevel;
-import com.microsoft.azure.cosmosdb.DocumentClientException;
-import com.microsoft.azure.cosmosdb.FeedOptions;
-import com.microsoft.azure.cosmosdb.SqlParameter;
-import com.microsoft.azure.cosmosdb.SqlParameterCollection;
-import com.microsoft.azure.cosmosdb.SqlQuerySpec;
-import com.microsoft.azure.cosmosdb.rx.AsyncDocumentClient;
-import com.microsoft.azure.cosmosdb.rx.TestConfigurations;
-import org.testng.annotations.Test;
-import rx.Observable;
-
-import java.util.Arrays;
-import java.util.List;
-
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertNotNull;
-import static org.testng.Assert.assertTrue;
-
-public class MapperTest {
-
-    private final static int LARGE_TIMEOUT = 30000;
-
-    private AsyncDocumentClient client = new AsyncDocumentClient.Builder()
-            .withServiceEndpoint(TestConfigurations.HOST)
-            .withMasterKey(TestConfigurations.MASTER_KEY)
-            .withConnectionPolicy(ConnectionPolicy.GetDefault())
-            .withConsistencyLevel(ConsistencyLevel.Session)
-            .build();
+public class MapperTest extends AbstractMapperClass {
 
     private Mapper<Person> mapper = MappingManager.of(client).mapper(Person.class);
 
 
-    @Test(groups = {"internal"}, timeOut = LARGE_TIMEOUT)
-    public void shouldCreateEntity() {
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        Observable<Person> observable = mapper.save(ada);
-
-        assertNotNull(observable);
-
-        Person person = observable.toBlocking().first();
-
-        assertNotNull(person);
-
-        assertEquals(ada.getName(), person.getName());
-        assertEquals(ada.getAge(), person.getAge());
-        assertEquals(ada.getId(), person.getId());
-
+    @Override
+    Mapper<Person> getMapper() {
+        return mapper;
     }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenCreateHasEntityNull() {
-        mapper.save((Person) null);
-    }
-
-    @Test(groups = {"internal"}, timeOut = LARGE_TIMEOUT)
-    public void shouldCreateEntities() {
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        Person marie = new Person();
-        marie.setId("marie" + System.currentTimeMillis());
-        marie.setName("Marie Curie");
-        marie.setAge(20);
-
-        Observable<List<Person>> observable = mapper.save(Arrays.asList(ada, marie));
-
-        assertNotNull(observable);
-
-        List<Person> people = observable.toBlocking().first();
-
-        assertNotNull(people);
-
-        assertTrue(people.stream().map(Person::getId)
-                .allMatch(id -> id.equals(ada.getId()) || id.equals(marie.getId())));
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenEntitiesIsNull() {
-        mapper.save((Iterable<Person>) null);
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenEntitiesHasANullElement() {
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        mapper.save(Arrays.asList(ada, null));
-    }
-
-
-    @Test(groups = {"internal"})
-    public void shouldUpdateEntity() {
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        mapper.save(ada).toBlocking().first();
-
-        ada.setName("Ada Lovelace update");
-
-        Person person = mapper.save(ada).toBlocking().first();
-
-        assertEquals(ada.getName(), person.getName());
-        assertEquals(ada.getId(), person.getId());
-        assertEquals(ada.getAge(), person.getAge());
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenFindByIdHasNullId() {
-        mapper.findById((String) null);
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenFindByIdHasNullIds() {
-        mapper.findById((Iterable<String>) null);
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenFindByIdHasANullElement() {
-
-        mapper.findById(Arrays.asList("id", null));
-    }
-
-    @Test(groups = {"internal"}, timeOut = LARGE_TIMEOUT)
-    public void shouldReturnErrorWhenIdDoesNotExist() {
-        Observable<Person> observable = mapper.findById("not_found");
-        try {
-            observable.single().toBlocking().first();
-        } catch (Exception ex) {
-            if (ex.getCause() instanceof DocumentClientException) {
-                return;
-            }
-        }
-
-        assert false;
-    }
-
-    @Test(groups = {"internal"}, timeOut = LARGE_TIMEOUT)
-    public void shouldFindById() {
-
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        mapper.save(ada).toBlocking().first();
-
-        Observable<Person> result = mapper.findById(ada.getId());
-
-        Person person = result.single().toBlocking().first();
-
-        assertEquals(ada.getName(), person.getName());
-        assertEquals(ada.getId(), person.getId());
-        assertEquals(ada.getAge(), person.getAge());
-    }
-
-    @Test(groups = {"internal"}, timeOut = LARGE_TIMEOUT)
-    public void shouldFindByIds() {
-
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        Person marie = new Person();
-        marie.setId("marie" + System.currentTimeMillis());
-        marie.setName("Marie Curie");
-        marie.setAge(20);
-
-        Observable<List<Person>> observable = mapper.save(Arrays.asList(ada, marie));
-
-        assertNotNull(observable);
-
-        observable.toBlocking().first();
-
-
-        Observable<List<Person>> result = mapper.findById(Arrays.asList(ada.getId(), marie.getId()));
-
-        List<Person> people = result.toBlocking().first();
-
-        assertNotNull(people);
-
-        assertTrue(people.stream().map(Person::getId)
-                .allMatch(id -> id.equals(ada.getId()) || id.equals(marie.getId())));
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenDeleteByIdHasNullId() {
-        mapper.deleteById((String) null);
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenDeleteByHasNullIds() {
-        mapper.deleteById((Iterable<String>) null);
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenDeleteByHasANullElement() {
-        mapper.deleteById(Arrays.asList("id", null));
-    }
-
-
-    @Test(groups = {"internal"})
-    public void shouldReturnErrorWhenDeleteDoesNotRemoveEntity() {
-        Observable<Void> observable = mapper.deleteById("not_found");
-        try {
-            observable.single().toBlocking().first();
-        } catch (Exception ex) {
-            if (ex.getCause() instanceof DocumentClientException) {
-                return;
-            }
-        }
-
-        assert false;
-    }
-
-    @Test(groups = {"internal"})
-    public void shouldRemoveById() {
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        mapper.save(ada).toBlocking().first();
-
-        Observable<Void> observable = mapper.deleteById(ada.getId());
-        observable.toBlocking().first();
-
-        try {
-            mapper.findById(ada.getId()).toBlocking().first();
-        } catch (RuntimeException ex) {
-            if (ex.getCause() instanceof DocumentClientException) {
-                return;
-            }
-        }
-
-        assert false;
-
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenQueryHasNullStringQuery() {
-
-        FeedOptions queryOptions = new FeedOptions();
-        mapper.query((String) null, queryOptions);
-    }
-
-    @Test(groups = {"internal"}, expectedExceptions = NullPointerException.class)
-    public void shouldReturnErrorWhenQueryHasNullQueryOptions() {
-        mapper.query("select * from c", null);
-    }
-
-    @Test(groups = {"internal"})
-    public void shouldExecuteQuery() {
-
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        Person marie = new Person();
-        marie.setId("marie" + System.currentTimeMillis());
-        marie.setName("Marie Curie");
-        marie.setAge(20);
-
-        mapper.save(Arrays.asList(ada, marie)).toCompletable().await();
-
-        FeedOptions queryOptions = new FeedOptions();
-        queryOptions.setMaxItemCount(100);
-
-        Observable<List<Person>> result =
-                mapper.query("SELECT * FROM Person WHERE Person.name = 'Ada Lovelace'", queryOptions);
-
-        List<Person> people = result.toBlocking().first();
-
-        assertTrue(people.size() > 1);
-        assertTrue(people.stream().map(Person::getName).allMatch(ada.getName()::equals));
-    }
-
-    @Test(groups = {"internal"})
-    public void shouldExecuteQueryBySpec() {
-
-        Person ada = new Person();
-        ada.setId("ada" + System.currentTimeMillis());
-        ada.setName("Ada Lovelace");
-        ada.setAge(20);
-
-        Person marie = new Person();
-        marie.setId("marie" + System.currentTimeMillis());
-        marie.setName("Marie Curie");
-        marie.setAge(20);
-
-        mapper.save(Arrays.asList(ada, marie)).toCompletable().await();
-
-        FeedOptions queryOptions = new FeedOptions();
-        queryOptions.setMaxItemCount(100);
-
-        SqlQuerySpec query = new SqlQuerySpec("SELECT * FROM Person WHERE Person.name = @name",
-                new SqlParameterCollection(new SqlParameter("@name", ada.getName())));
-
-        Observable<List<Person>> result = mapper.query(query, queryOptions);
-
-        List<Person> people = result.toBlocking().first();
-
-        assertTrue(people.size() > 1);
-        assertTrue(people.stream().map(Person::getName).allMatch(ada.getName()::equals));
-    }
-
 }
